🎉 feat(customer): add existsPersonWithId method
Adds the existsPersonWithId method to both JPA and list data access services
to check if the customer ID exists before deleting it. Also, throws a
ResourceNotFoundException if the customer does not exist.

+++ src/main/java/com/kakreak/customer/CustomerService.java
@@ -41,6 +41,12 @@ public class CustomerService {
     }
 
     public void deleteCustomerById(Integer customerId) {
+
+        // check if customerID exist
+        if (!customerDao.exitsPersonWithId(customerId)) {
+            throw new ResourceNotFoundException("Customer " + customerId + " does not exist.");
+        }
+
         customerDao.deleteCustomer(customerId);
     }
 }
